api/requirements.txt
alembic==1.7.7
black==22.3.0
blinker==1.4
click==8.1.3
Flask==2.1.2
Flask-HTTPAuth==4.6.0
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
SQLAlchemy==1.4.36
tomli==2.0.1
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
